splitting files up
